Patch `QualifiedName`’s `namespaceIndex` data size
According to the spec, the `namespaceIndex` should be an UInt16, not an
Int as the xml document specifies.

To simplify the process of applying patches—and to future proof—I’ve
added support for applying multiple patches and grouped the patches by
version.

+++ stack/uatype/uatype_auto.go
@@ -415,7 +415,7 @@ type DiagnosticInfo struct {
 
 // QualifiedName is a string qualified with a namespace index.
 type QualifiedName struct {
-	NamespaceIndex int32
+	NamespaceIndex uint16
 	Name           string
 }
